+++ Controllers/PreferentialsController.cs
@@ -1,3 +1,5 @@
+using System.IdentityModel.Tokens.Jwt;
+using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
@@ -6,6 +8,7 @@ using Microsoft.AspNetCore.Mvc;
 using WorkAppReactAPI.Assets;
 using WorkAppReactAPI.Configuration;
 using WorkAppReactAPI.Data.Interface;
+using WorkAppReactAPI.Dtos;
 using WorkAppReactAPI.Dtos.Requests;
 
 namespace WorkAppReactAPI.Controllers
@@ -28,20 +31,20 @@ namespace WorkAppReactAPI.Controllers
 
 
         [HttpGet]
-        public async Task<ActionResult<DynamicResult>> getAllPreferential()
+        public async Task<ActionResult<DynamicResult>> getAllPreferential([FromQuery] Query model)
         {
 
-            var result = await _repository.ListPreferentials();
-            return result;
+            var result = await _repository.ListPreferentials(model);
+            return Ok(result);
         }
 
         [HttpPost]
         [Route("add")]
-        public async Task<ActionResult<DynamicResult>> AddPreferential([FromForm] PreferentialUpdate model)
+        public async Task<ActionResult<DynamicResult>> AddPreferential([FromForm] PreferentialUpdate model, [FromHeader] HeaderParamaters header)
         {
 
             var file = model.Image;
-            if (file.Length > 0)
+            if (file != null)
             {
                 var path = _hostingEnvironment.UploadImage(file, "\\Upload\\Images\\");
                 if (path.Length == 0)
@@ -54,21 +57,33 @@ namespace WorkAppReactAPI.Controllers
                 }
                 model.ImageUrl = path;
             }
-            var result = await _repository.AddPreferential(model);
-            if (result.Status != 1)
+              var handler = new JwtSecurityTokenHandler();
+            var tokenStr = header.Authorization.Substring("Bearer ".Length).Trim();
+            var jsonToken = handler.ReadToken(tokenStr);
+            var tokenS = jsonToken as JwtSecurityToken;
+
+            var auth = new UserLogin()
+            {
+                Phone = tokenS.Claims.First(claim => claim.Type == "Phone").Value,
+                Password = tokenS.Claims.First(claim => claim.Type == "Password").Value,
+                isCustomer = bool.Parse(tokenS.Claims.First(claim => claim.Type == "isCustomer").Value),
+                Role = int.Parse(tokenS.Claims.First(claim => claim.Type == "Role").Value)
+            };
+            var result = await _repository.AddPreferential(model, auth);
+            if (file != null && result.Status != 1)
             {
                 System.IO.File.Delete(model.ImageUrl);
             }
-            return result;
+            return Ok(result);
         }
 
         [HttpPost]
         [Route("update")]
-        public async Task<ActionResult<DynamicResult>> UpdatePreferential([FromBody] PreferentialUpdate model)
+        public async Task<ActionResult<DynamicResult>> UpdatePreferential([FromQuery]string code,[FromForm] PreferentialUpdate model, [FromHeader] HeaderParamaters header)
         {
 
             var file = model.Image;
-            if (file.Length > 0)
+            if (file != null)
             {
                 var path = _hostingEnvironment.UploadImage(file, "\\Upload\\Images\\");
                 if (path.Length == 0)
@@ -81,12 +96,43 @@ namespace WorkAppReactAPI.Controllers
                 }
                 model.ImageUrl = path;
             }
-            var result = await _repository.UpdatePreferential(model);
-            if (result.Status != 1)
+             var handler = new JwtSecurityTokenHandler();
+            var tokenStr = header.Authorization.Substring("Bearer ".Length).Trim();
+            var jsonToken = handler.ReadToken(tokenStr);
+            var tokenS = jsonToken as JwtSecurityToken;
+
+            var auth = new UserLogin()
+            {
+                Phone = tokenS.Claims.First(claim => claim.Type == "Phone").Value,
+                Password = tokenS.Claims.First(claim => claim.Type == "Password").Value,
+                isCustomer = bool.Parse(tokenS.Claims.First(claim => claim.Type == "isCustomer").Value),
+                Role = int.Parse(tokenS.Claims.First(claim => claim.Type == "Role").Value)
+            };
+            var result = await _repository.UpdatePreferential(code , model, auth);
+            if (file != null && result.Status != 1)
             {
                 System.IO.File.Delete(model.ImageUrl);
             }
 
+            return Ok(result);
+        }
+        [HttpPost]
+        [Route("delete")]
+        public async Task<ActionResult<DynamicResult>> DeletePreferential([FromQuery]string code,[FromHeader] HeaderParamaters header)
+        {
+            var handler = new JwtSecurityTokenHandler();
+            var tokenStr = header.Authorization.Substring("Bearer ".Length).Trim();
+            var jsonToken = handler.ReadToken(tokenStr);
+            var tokenS = jsonToken as JwtSecurityToken;
+
+            var auth = new UserLogin()
+            {
+                Phone = tokenS.Claims.First(claim => claim.Type == "Phone").Value,
+                Password = tokenS.Claims.First(claim => claim.Type == "Password").Value,
+                isCustomer = bool.Parse(tokenS.Claims.First(claim => claim.Type == "isCustomer").Value),
+                Role = int.Parse(tokenS.Claims.First(claim => claim.Type == "Role").Value)
+            };
+            var result = await _repository.DeletePreferential(code, auth);
             return result;
         }
